fix: add back separate queue for tx monitoring
