+++ src/constants.ts
@@ -1,4 +1,7 @@
 export namespace ReconnectionServiceConstants {
   export const GRAPH_UPDATE_QUEUE_NAME = 'graphUpdateQueue';
   export const GRAPH_UPDATE_JOB_TYPE = 'graphUpdate';
+  export const TX_MONITOR_QUEUE_NAME = 'graphTxMonitorQueue';
+  export const TX_MONITOR_JOB_NAME = 'graphTxMonitorJob';
+  export const EPOCH_CAPACITY_PREFIX = 'epochCapacity:';
 }
