+++ src/processor/processor.module.ts
@@ -12,13 +12,14 @@ import { BullBoardModule } from '@bull-board/nestjs';
 import { BullMQAdapter } from '@bull-board/api/bullMQAdapter';
 import { ExpressAdapter } from '@bull-board/express';
 import { ReconnectionServiceConstants } from '#app/constants';
-import { QueueConsumerService } from './queue-consumer.service';
+import { GraphUpdateQueueConsumerService } from './graph-update-queue-consumer.service';
 import { GraphUpdateCompletionMonitorService } from './graph-update-completion-monitor.service';
 import { ReconnectionGraphService } from './reconnection-graph.service';
 import { GraphManagerModule } from '../graph/graph-state.module';
 import { GraphStateManager } from '../graph/graph-state-manager';
 import { ProviderWebhookService } from './provider-webhook.service';
 import { NonceService } from './nonce.service';
+import { TxMonitorQueueConsumerService } from './tx-monitor-queue-consumer.service';
 
 @Module({
   imports: [
@@ -57,6 +58,17 @@ import { NonceService } from './nonce.service';
         removeOnFail: false,
       },
     }),
+    BullModule.registerQueue({
+      name: ReconnectionServiceConstants.TX_MONITOR_QUEUE_NAME,
+      defaultJobOptions: {
+        attempts: 3,
+        backoff: {
+          type: 'exponential',
+        },
+        removeOnComplete: true,
+        removeOnFail: true,
+      },
+    }),
     // Bullboard UI
     BullBoardModule.forRoot({
       route: '/queues',
@@ -66,12 +78,24 @@ import { NonceService } from './nonce.service';
       name: ReconnectionServiceConstants.GRAPH_UPDATE_QUEUE_NAME,
       adapter: BullMQAdapter,
     }),
+    BullBoardModule.forFeature({
+      name: ReconnectionServiceConstants.TX_MONITOR_QUEUE_NAME,
+      adapter: BullMQAdapter,
+    }),
     ConfigModule,
     GraphManagerModule,
     BlockchainModule,
   ],
   controllers: [],
-  providers: [QueueConsumerService, GraphUpdateCompletionMonitorService, ReconnectionGraphService, GraphStateManager, ProviderWebhookService, NonceService],
+  providers: [
+    GraphUpdateQueueConsumerService,
+    TxMonitorQueueConsumerService,
+    GraphUpdateCompletionMonitorService,
+    ReconnectionGraphService,
+    GraphStateManager,
+    ProviderWebhookService,
+    NonceService,
+  ],
   exports: [ReconnectionGraphService, BullModule],
 })
 export class ProcessorModule {}
